chat

+++ SEP3-tier2/Models/SaleOffer.cs
@@ -2,9 +2,14 @@
 {
     public class SaleOffer
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public long item_id { get; set; }
         public int sale_price { get; set; }
         public long wallet_id { get; set; }
     }
+    
+    
+    
+    
+    
 }
